+++ packages/sdk-ink/src/get-contract.ts
@@ -4,6 +4,7 @@ import {
   wrapAsyncTx,
 } from "@polkadot-api/common-sdk-utils"
 import type { InkClient, InkMetadataLookup } from "@polkadot-api/ink-contracts"
+import { Binary, compactNumber } from "@polkadot-api/substrate-bindings"
 import { Enum, SS58String } from "polkadot-api"
 import type {
   GenericInkDescriptors,
@@ -72,23 +73,57 @@ export function getContract<
         data,
       )
       if (response.result.success) {
+        const availableData = {
+          events: inkClient.event.filter(mapAddr(address), response.events),
+          gasRequired: response.gas_required,
+          storageDeposit: getSignedStorage(response.storage_deposit),
+          send: () => {
+            const limitParams = {
+              gasLimit: response.gas_required,
+              storageDepositLimit: getStorageLimit(response.storage_deposit),
+            }
+            return contractApi.send(message, {
+              ...args,
+              ...limitParams,
+            })
+          },
+        }
+
+        // In case REVERT flag is set, it might return a string with the error message or debug info.
+        if (response.result.value.flags & 0x01) {
+          const data = response.result.value.data
+          // In case of panic! or return_value(REVERT, &"some message"), the value comes back as an opaque string
+          // Meaning what we get over the wire is [payload_len,data]
+          // And `data` is a Vec<u8>, hence [msg_len,...chars]. In this case, the msg_len is redundant.
+          const decodeMessage = () => {
+            try {
+              const bytes = data.asBytes()
+              const length = compactNumber.dec(bytes)
+              const compactLength = compactNumber.enc(length).length
+              if (compactLength + length === bytes.length) {
+                return Binary.fromBytes(bytes.slice(compactLength)).asText()
+              }
+            } catch {}
+            return data.asHex()
+          }
+          return {
+            success: false,
+            value: {
+              type: "FlagReverted",
+              value: {
+                ...availableData,
+                message: decodeMessage(),
+                raw: response.result.value.data,
+              },
+            },
+          }
+        }
+
         const decoded = msg.decode(response.result.value)
         return mapResult(flattenResult(decoded), {
           value: (innerResponse) => ({
             response: innerResponse,
-            events: inkClient.event.filter(mapAddr(address), response.events),
-            gasRequired: response.gas_required,
-            storageDeposit: getSignedStorage(response.storage_deposit),
-            send: () => {
-              const limitParams = {
-                gasLimit: response.gas_required,
-                storageDepositLimit: getStorageLimit(response.storage_deposit),
-              }
-              return contractApi.send(message, {
-                ...args,
-                ...limitParams,
-              })
-            },
+            ...availableData,
           }),
         })
       }